Local time conversion
Added a simple UTC timestamp to local time converter. It handles leap years
until year 2100, time zones, and DST (only month/week/day start/end times
are currently supported).

This will be used on the Nixie clock to display the correct time from the
GPS timestamp, and possibly to display the date in the future.

The tests folder contains tests for the local time converter, that can be
run by launching “make test” from the folder.