Integrate date/time calculation into the main code
Timer0 interrupt is used to maintain a tick count which approximates wallclock
time.
The main code periodically check if the tick count changed, and if that’s the
case, converts it to a timestamp then to the local date/time.

The code is structured to correctly work in the event that the main code would
take multiple ticks to fully execute. This should hopefully not happen in
practice. (For instance, the date/time conversion code can be executed ~600
times per second, and there are ~10.5 ticks per second, so there is room for
now)