Improve GPS initialization sequence
When the serial port reception is enabled, the GPS might be in the process
of sending a message. To ensure proper synchronization, the following sequence
is now used:
 * Initialize the GPS to put it in a mode where no messages are sent (except
   possibly some acknowledgment messages)
 * Wait a bit so all messages that the GPS may have queued up have been sent
 * Enable serial port reception
 * Ask the GPS to periodically send message 7 (clock message).

This guarantees that the first message received will be a full acknowledgment
message (for the clock message enable command), followed by a full clock
message, periodically.