Fix GPS initialization, again
The previous code worked during debugging, but not on cold power on, because
the GPS takes some time to start up, and require a delay between messages.

Since the GPS mode is switched at startup, and only sends a message every 10
seconds nominally, it’s not easy to detect if the GPS is already configured
or not, so the code only sends commands in sequence with some delays between
them. Initialization is considered complete when one message 7 (clock) is
received.