Synchronize the timer with the GPS
This actually makes the clock display the current time, once the GPS is
synchronized.

GPS calculation are made using 64-bit variables, even though the result can
(and should) fit into lower sized variables.
The problem (I think) is that all multiplication/divisions are translated
by the compiler to function calls, and the called function needs the arguments
and the return value to be the same size. Unfortunately, the compiler only
seems to consider the return value to pick the argument size, so if the
return value is 8-bit, the dividend and divisor will be cast to 8-bit before
being divided. This will obviously return incorrect values.

The code sets the number of ticks to the current GPS time, but also set the
timer value so its next tick will be synchronized (roughly) to the actual
GPS time. This results in better precision, but is also important since the
timer value can result in ticks “ahead of time” from the GPS time, which will
result in the clock occasionally briefly going back to the previous second
(because a tick incremented the tick count early, then the GPS set the tick
count back one tick).