Sleep support
Since the clock is driven by interrupts (from timer or serial) and does nothing
except checking if a tick happened outside of an interrupt, put the PIC to
sleep before waiting for the next interrupt.

The code may present a race condition if an interrupt happens just before
going to sleep; this should not happen in the normal case because the full
handling of any interrupt will take less time that the delay between two
timer interrupts.