Better GPS error/sync management
* Error status is separated from sync status; the Nixie clock digits separators
are used to report the error status (left separator not blinking = error) and
the sync status (right separator not blinking = not synchronized). The actual
error code, if any, is still displayed on the status LED.

* The GPS now continues processing messages if an error occurs, but doesn’t
reset the error status right away; it’s only reset when 255 valid messages
have been received from the GPS (which will take around 42 minutes), so the
error will be visible by the user.
* Error mode is separated from sync stat